change external group type to i64; use SD seed as external group; fix some sql, html and css

+++ src/model/read.rs
@@ -40,7 +40,7 @@ pub struct Element {
     pub group_id: Option<u32>,
     /// Group info derived from external source
     #[sqlx(rename = "ext_group")]
-    pub group: Option<u32>,
+    pub group: Option<i64>,
 }   
 
 /// Element waiting for metadata download/parse
